Add manual Ruby install script; reconfigure Salt
I was getting frustrated trying to configure Salt to install and
configure Ruby for the vagrant user so that it could be used as the
vagrant user. While I've kept the rbenv state file in place, I've
removed it from being called in the top.sls. Instead the user now will
need to call the post-salt-install/rbenv.sh script to install Ruby after
the core packages are installed through Salt.

This isn't perfect, but there are some advantages to having the Ruby
install as a separate script rather than something that Salt configures
automatically.

I need to write a README now.