Update README; add gitignore
I didn't like having yet another directory attached to this build
process, so I tucked the `post-salt-install/` directory into the `salt/`
directory to clean up the footprint a little. I updated the README to
address that change.

Added the fact that we're doing stuff with NodeJS for this build as well
so that people do not need to include therubyracer gem when working with
Rails.